kinda awkward lazy load qml
qml is fairly expensive so we let's not load it until the user activates
the KSNI. this is achieved by managing the entire qml shebang's lifetime
on-demand via a trivial wrapper object. this allows us to end the
lifetime from qml (effectively a glorified window hiding) which has the
advantage that it also brings down the footprint a bit when the window
goes away but the app wasn't quit (e.g. it's only acting as
informational display) and also means we'll not have to negotiate which
window is the main window between the cpp and qml side.