always start nomodeset
in lieu of plasma getting managed by systemd we'll want this always up
when kglobalaccesl is up (so long as the nomodeset kernel arg is set
anyway).
this may mess with neon's oem mode, needs investigating. at the same
time I'm not sure why oem mode would run with nomodeset