add a cpp property to track whether nomodeset perists
neon will also set the resulting system to nomodeset by default when the
ISO was started in nomodeset mode. since this is somewhat specific to
neon we'll only want to show an additional hint about this when that
property is true.

trivial check for this is running through os-release data